code review

+++ src/net-questdb-client-tests/QuestDbQueryIntegrationTests.cs
@@ -247,7 +247,7 @@ public class QuestDbQueryIntegrationTests
         }
 
         public override void OnEnd(long totalRows) => Ended = true;
-        public override void OnExecDone(short opType, long rowsAffected) => ExecDoneObserved = true;
+        public override void OnExecDone(byte opType, long rowsAffected) => ExecDoneObserved = true;
         public override void OnError(byte status, string message) => LastErrorStatus = status;
 
         public long[] AllLongs(int colIndex)
